Better separation of responsibility
