+++ GrpcSaga.Api/Services/CartoonApiService.cs
@@ -1,42 +1,20 @@
 ﻿using Cartoonalogue.Api.ViewModels;
-using CartoonDomain.Shared.Queries.v1.Contracts;
-using CartoonDomain.Shared.Commands.v1.Contracts;
-using CartoonDomain.Shared.v1.Interfaces;
-using StudioDomain.Shared.Queries.v1.Interfaces;
-using StudioDomain.Shared.Queries.v1.Contracts;
-using StudioDomain.Shared.Commands.v1.Interfaces;
-using StudioDomain.Shared.Commands.v1.Contracts;
-using CartoonDomain.Shared.Commands.v1.Interfaces;
-using Grpc.Net.Client;
-using ProtoBuf.Grpc.Client;
+
 
 namespace Cartoonalogue.Api.Services;
 
 public class CartoonApiService : ICartoonApiService
 {
     private readonly ILogger<CartoonApiService> _logger;
-    private readonly ICartoonDomainQueryService _cartoonDomainQueryService;
-    private readonly IStudioDomainQueryService _studioDomainQueryService;
-    private readonly ICartoonDomainCommandService _cartoonDomainCommandService;
-    private readonly IStudioDomainCommandService _studioDomainCommandService;
+    private readonly IDataService _dataService;
 
-    public CartoonApiService(ILogger<CartoonApiService> logger)
+
+    public CartoonApiService(ILogger<CartoonApiService> logger, IDataService dataService)
     {
         _logger = logger;
-
-        var CartoonQueryChannel = GrpcChannel.ForAddress("https://localhost:7227");
-        _cartoonDomainQueryService = CartoonQueryChannel.CreateGrpcService<ICartoonDomainQueryService>();
-
-        var CartoonCommandChannel = GrpcChannel.ForAddress("https://localhost:7006");
-        _cartoonDomainCommandService = CartoonCommandChannel.CreateGrpcService<ICartoonDomainCommandService>();
-
-        var studioQueryChannel = GrpcChannel.ForAddress("https://localhost:7129");
-        _studioDomainQueryService = studioQueryChannel.CreateGrpcService<IStudioDomainQueryService>();
-
-        var studioCommandChannel = GrpcChannel.ForAddress("https://localhost:7166");
-        _studioDomainCommandService = studioCommandChannel.CreateGrpcService<IStudioDomainCommandService>();
+        _dataService = dataService;
     }
-    
+
     public async Task<CartoonDetailsViewModel>? CreateCartoonDetailsAsync(CartoonDetailsCreateViewModel createViewModel)
     {
         if (createViewModel == null)
@@ -47,54 +25,8 @@ public class CartoonApiService : ICartoonApiService
         }
         try
         {
-            var isNewStudio = false;
-
-            // 1. Either insert or retrieve the studio
-            StudioViewModel? studioViewModel;
-            if (createViewModel.Cartoon.StudioId <= 0)
-            {
-                // Add new Studio
-                studioViewModel = await InsertStudioAsync(createViewModel.Studio);
-                if (studioViewModel == null)
-                {
-                    // Studio database update failed
-                    return null;
-                }
-                isNewStudio = true;
-            }
-            else
-            {
-                // Studio already exists so retrieve it
-                var studioRequest = new StudioRequest
-                {
-                    Id = createViewModel.Cartoon.StudioId
-                };
-                var studioResponse = await _studioDomainQueryService.GetStudioByIdAsync(studioRequest);
-                studioViewModel = new StudioViewModel
-                {
-                    Id = studioResponse.Id,
-                    Name = studioResponse.Name
-                };
-            }
-
-            // 2. Insert Cartoon
-            var cartoonViewModel = await InsertCartoonDetailsAsync(createViewModel, studioViewModel.Id);
-            if (cartoonViewModel == null)
-            {
-                // Cartoon database update failed
-                if (isNewStudio)
-                {
-                    await CompensateStudioAsync(studioViewModel);
-                }
-                return null;
-            }
-
-            var response = new CartoonDetailsViewModel
-            {
-                Cartoon = cartoonViewModel,
-                Studio = studioViewModel
-            };
-            return response;
+            var viewModel = await _dataService.CreateCartoonDetailsAsync(createViewModel);
+            return viewModel;
         }
         catch (Exception ex)
         {
@@ -116,31 +48,8 @@ public class CartoonApiService : ICartoonApiService
         }
         try
         {
-            var createRequest = new CartoonCreateRequest
-            {
-                Title = createViewModel.Title,
-                YearBegin = createViewModel.YearBegin,
-                YearEnd = createViewModel.YearEnd,
-                Rating = createViewModel.Rating,
-                Description = createViewModel.Description,
-                StudioId = createViewModel.StudioId
-            };
-            var createResponse = await _cartoonDomainCommandService.CreateCartoonAsync(createRequest);
-            if (createResponse == null)
-            {
-                return null;
-            }
-            var responseViewModel = new CartoonViewModel
-            {
-                Id = createResponse.Id,
-                Title = createResponse.Title,
-                YearBegin = createResponse.YearBegin,
-                YearEnd = createResponse.YearEnd,
-                Rating = createResponse.Rating,
-                Description = createResponse.Description,
-                StudioId = createResponse.StudioId
-            };
-            return responseViewModel;
+            var viewModel = await _dataService.CreateCartoonAsync(createViewModel);
+            return viewModel;
         }
         catch (Exception ex)
         {
@@ -163,25 +72,8 @@ public class CartoonApiService : ICartoonApiService
 
         try
         {
-            var createRequest = new CharacterCreateRequest
-            {
-                Name = requestViewModel.Name,
-                Description = requestViewModel.Description,
-                CartoonId = requestViewModel.CartoonId
-            };
-            var createResponse = await _cartoonDomainCommandService.CreateCharacterAsync(createRequest);
-            if (createResponse == null)
-            {
-                return null;
-            }
-            var responseViewModel = new CharacterViewModel
-            {
-                Id = createResponse.Id,
-                Name = createResponse.Name,
-                Description = createResponse.Description,
-                CartoonId = createResponse.CartoonId
-            };
-            return responseViewModel;
+            var viewModel = await _dataService.CreateCharacterAsync(requestViewModel);
+            return viewModel;
         }
         catch (Exception ex)
         {
@@ -202,21 +94,8 @@ public class CartoonApiService : ICartoonApiService
 
         try
         {
-            var createRequest = new StudioCreateRequest
-            {
-                Name = createViewModel.Name
-            };
-            var createResponse = await _studioDomainCommandService.CreateStudioAsync(createRequest);
-            if (createResponse == null)
-            {
-                return null;
-            }
-            var studioViewModel = new StudioViewModel
-            {
-                Id = createResponse.Id,
-                Name = createResponse.Name
-            };
-            return studioViewModel;
+            var viewModel = await _dataService.CreateStudioAsync(createViewModel);
+            return viewModel;
         }
         catch (Exception ex)
         {
@@ -239,27 +118,8 @@ public class CartoonApiService : ICartoonApiService
         try
         {
 
-            var updateRequest = new CartoonUpdateRequest
-            {
-                Id = updateViewModel.Id,
-                Title = updateViewModel.Title,
-                YearBegin = updateViewModel.YearBegin,
-                YearEnd = updateViewModel.YearEnd,
-                Description = updateViewModel.Description,
-                Rating = updateViewModel.Rating,
-                StudioId = updateViewModel.StudioId
-            };
-            var updateResponse = await _cartoonDomainCommandService.UpdateCartoonAsync(updateRequest);
-
-            var cartoonRequest = new CartoonRequest
-            {
-                Id = updateViewModel.Id
-            };
-
-            var viewModelResponse = await _cartoonDomainQueryService.GetCartoonByIdAsync(cartoonRequest);
-            var cartoonViewModel = await MapCartoonAsync(viewModelResponse);
-
-            return cartoonViewModel;
+            var viewModel = await _dataService.UpdateCartoonAsync(updateViewModel);
+            return viewModel;
         }
         catch (Exception ex)
         {
@@ -276,31 +136,7 @@ public class CartoonApiService : ICartoonApiService
     {
         try
         {
-            var cartoonRequest = new CartoonRequest
-            {
-                Id = id
-            };
-
-            var cartoonResponse = await _cartoonDomainQueryService.GetCartoonByIdAsync(cartoonRequest);
-
-            if (cartoonResponse == null)
-            {
-                // The requested id is not found
-                var correlationId = Guid.NewGuid();
-                // Write Not Found, requested Id, and correlation Id to log
-                return null;
-            }
-
-            var viewModel = new CartoonViewModel
-            {
-                Id = cartoonResponse.Id,
-                Title = cartoonResponse.Title,
-                Description = cartoonResponse.Description,
-                YearBegin = cartoonResponse.YearBegin,
-                YearEnd = cartoonResponse.YearEnd,
-                Rating = cartoonResponse.Rating,
-                StudioId = cartoonResponse.StudioId
-            };
+            var viewModel = await _dataService.GetCartoonByIdAsync(id);
 
             return viewModel;
         }
@@ -319,24 +155,8 @@ public class CartoonApiService : ICartoonApiService
     {
         try
         {
-            var cartoonRequest = new CartoonRequest
-            {
-                Id = id
-            };
-
-            var cartoonResponse = await _cartoonDomainQueryService.GetCartoonByIdAsync(cartoonRequest);
-
-            if (cartoonResponse == null)
-            {
-                // The requested id is not found
-                var correlationId = Guid.NewGuid();
-                // Write Not Found, requested Id, and correlation Id to log
-                return null;
-            }
-
-            var viewModelResponse = await MapCartoonDetailsAsync(cartoonResponse);
-
-            return viewModelResponse;
+            var viewModel = await _dataService.GetCartoonDetailsByIdAsync(id);
+            return viewModel;
         }
         catch (Exception ex)
         {
@@ -351,25 +171,9 @@ public class CartoonApiService : ICartoonApiService
 
     public async Task<IList<CartoonViewModel>> GetCartoonListAsync()
     {
-        // TODO: Change this to just cartoon, not info
         try
         {
-            var getShowResponse = await _cartoonDomainQueryService.GetAllCartoonsAsync();
-
-            if (getShowResponse == null || getShowResponse.Cartoons.Count() == 0)
-            {
-                // The requested id is not found
-                var correlationId = Guid.NewGuid();
-                // Write Not Found, requested Id, and correlation Id to log
-                return null;
-            }
-
-            var cartoons = new List<CartoonViewModel>();
-            foreach (var c in getShowResponse.Cartoons)
-            {
-                var currentCartoon = await MapCartoonAsync(c);
-                cartoons.Add(currentCartoon);
-            }
+            var cartoons = await _dataService.GetCartoonListAsync();
             return cartoons;
         }
         catch (Exception ex)
@@ -382,169 +186,5 @@ public class CartoonApiService : ICartoonApiService
             throw;
         }
     }
-
-    private async Task<CartoonViewModel> MapCartoonAsync(CartoonResponse response)
-    {
-        var currentCartoon = new CartoonViewModel
-        {
-            Id = response.Id,
-            Title = response.Title,
-            YearBegin = response.YearBegin,
-            YearEnd = response.YearEnd,
-            Description = response.Description,
-            Rating = response.Rating,
-            StudioId = response.StudioId
-        };
-        return currentCartoon;
-    }
-
-    private async Task<CartoonDetailsViewModel> MapCartoonDetailsAsync(CartoonResponse response)
-    {
-        var cartoonDetails = new CartoonDetailsViewModel
-        {
-            Cartoon = new CartoonViewModel
-            {
-                Id = response.Id,
-                Title = response.Title,
-                YearBegin = response.YearBegin,
-                YearEnd = response.YearEnd,
-                Description = response.Description,
-                Rating = response.Rating,
-                StudioId= response.StudioId
-            },
-            Studio = new StudioViewModel
-            {
-                Id = response.StudioId,
-                Name = (await _studioDomainQueryService.GetStudioByIdAsync(new StudioRequest { Id = response.StudioId })).Name
-            }
-        };
-        if (response.Characters.Any())
-        {
-            cartoonDetails.Cartoon.Characters = new List<CharacterViewModel>();
-            foreach (var c in response.Characters)
-            {
-                cartoonDetails.Cartoon.Characters.Add(new CharacterViewModel
-                {
-                    Id = c.Id,
-                    Name = c.Name,
-                    Description = c.Description,
-                    CartoonId = c.CartoonId
-                });
-            }
-        }
-        return cartoonDetails;
-    }
-
-    private async Task<StudioViewModel?> InsertStudioAsync(StudioCreateViewModel studio)
-    {
-        var studioRequest = new StudioCreateRequest
-        {
-            Name = studio.Name
-        };
-        var studioResponse = await _studioDomainCommandService.CreateStudioAsync(studioRequest);
-        if (studioResponse == null)
-        {
-            // Studio insert failed
-            return null;
-        }
-        var response = new StudioViewModel
-        {
-            Id = studioResponse.Id,
-            Name = studioResponse.Name
-        };
-        return response;
-    }
-
-    private async Task<bool> CompensateStudioAsync(StudioViewModel studio)
-    {
-        var studioRequest = new StudioDeleteRequest
-        {
-            Id = studio.Id
-        };
-        var studioResponse = await _studioDomainCommandService.DeleteStudioAsync(studioRequest);
-        if (studioResponse == null)
-        {
-            // Log the delete failure
-        }
-        return studioResponse != null;
-    }
-
-    private async Task<CartoonViewModel?> InsertCartoonDetailsAsync(CartoonDetailsCreateViewModel createViewModel, int studioId)
-    {
-        var charactersCreateRequest = new List<CharacterCreateRequest>();
-        if (createViewModel.Characters != null && createViewModel.Characters.Any())
-        {
-            foreach (var c in createViewModel.Characters)
-            {
-                charactersCreateRequest.Add(new CharacterCreateRequest
-                {
-                    Name = c.Name,
-                    Description = c.Description
-                });
-            }
-        }
-        var cartoonRequest = new CartoonDetailsCreateRequest
-        {
-            Cartoon = new CartoonCreateRequest
-            {
-                Title = createViewModel.Cartoon.Title,
-                Description = createViewModel.Cartoon.Description,
-                YearBegin = createViewModel.Cartoon.YearBegin,
-                YearEnd = createViewModel.Cartoon.YearEnd,
-                Rating = createViewModel.Cartoon.Rating,
-                StudioId = studioId,
-                Characters = charactersCreateRequest
-            }
-        };
-
-        var cartoonResponse = await _cartoonDomainCommandService.CreateCartoonDetailsAsync(cartoonRequest);
-        if (cartoonResponse == null)
-        {
-            // Cartoon insert failed
-            return null;
-        }
-
-        var charactersResponse = new List<CharacterViewModel>();
-        if (cartoonResponse.Cartoon.Characters != null && cartoonResponse.Cartoon.Characters.Count > 0)
-        {
-            foreach (var c in cartoonResponse.Cartoon.Characters)
-            {
-                charactersResponse.Add(new CharacterViewModel
-                {
-                    Id = c.Id,
-                    Name = c.Name,
-                    Description = c.Description,
-                    CartoonId = c.CartoonId
-                });
-            }
-        }
-
-        var response = new CartoonViewModel
-        {
-            Id = cartoonResponse.Cartoon.Id,
-            Title = cartoonResponse.Cartoon.Title,
-            Description = cartoonResponse.Cartoon.Description,
-            YearBegin = cartoonResponse.Cartoon.YearBegin,
-            YearEnd = cartoonResponse.Cartoon.YearEnd,
-            Rating = cartoonResponse.Cartoon.Rating,
-            StudioId = cartoonResponse.Cartoon.StudioId,
-            Characters = charactersResponse
-        };
-
-        return response;
-    }
-
-    private async Task<List<CharacterViewModel?>> InsertCharacterListAsync(List<CharacterCreateViewModel> characters, CartoonViewModel cartoon)
-    {
-        return null;
-    }
-
-    private async Task DeleteStudioAsync(int id)
-    {
-        var request = new StudioDeleteRequest
-        {
-            Id = id
-        };
-        var response = await _studioDomainCommandService.DeleteStudioAsync(request);
-    }
 }
+
